+++ js-web/vue-vite/cypress/web-template/pages/layoutPublic.js
@@ -1,1 +1,0 @@
-export const layoutPublic = () => cy.get('[data-cy="layout-public"]')