+++ js-web/vue-vite/playwright.config.js
@@ -1,0 +1,10 @@
+const config = {
+  use: {
+    baseURL: 'http://localhost:8080',
+    browserName: 'chromium',
+    // screenshot: 'only-on-failure', // on, off, only-on-failure
+    headless: true
+  }
+}
+
+module.exports = config